Hotel Level Update
Hello Everyone, hopefully nothing within this update conflict with what has been added in the last 13 pulls. Mainly it adds more walls and more things in general. I had not added collision to the doorways or stairs yet and I fully realized I need to come up with a flat square wall for one of the hotel parts. But hopefully this update works out.