Minor Interactable changes
Fixed a bug with the inventory not working properly. Also added an interact button so the player can look at the object again. Press X button to back out from interact.
Also added a box collider to the note on the receptionist desk and had to disable collider for desk. Must find fix for this soon.